cleaned up demo files, removed some unused state properties.
