+++ demo/DragThingsToBoxes/Boxable.jsx
@@ -1,0 +1,26 @@
+import React from 'react';
+var shortid = require('shortid');
+import { DragDropContainer } from '../../src';
+
+/*
+    Boxable -- a thing you can drag into a Box
+*/
+
+export default class Boxable extends React.Component {
+    render() {
+      return (
+        <DragDropContainer
+          targetKey={this.props.targetKey}
+          dragData={{label: this.props.label}}
+          customDragElement={this.props.customDragElement}
+          onDragStart={()=>(console.log('start'))}
+          onDrag={()=>(console.log('dragging'))}
+          onDragEnd={()=>(console.log('end'))}
+          onDrop={(e)=>(console.log(e))}
+  
+        >
+          <img src={this.props.image} height="45" style={{ marginLeft: 40}}/>
+        </DragDropContainer>
+      );
+    }
+  }
